V1.0
Blender tool designed to convert the textures of any model in your scene to a retro, low-resolution style, emulating the look of textures used in classic consoles like the PS1/N64. With this tool, you can convert textures from models in your scene to a 16-color palette and simulate the graphical limitations of 90s-era gaming hardware.